Simplify release script and skip existing releases
We simplified the script and utilising better the `chart-releaser` CLI. Now we iterate through
all charts and we package them but we will upload only the packages are new. Added a new flag
for `chart-releaser` which will help us to skip the existing packages.

Ticket: https://mattermost.atlassian.net/browse/CLD-3970